// api/queries/cardPool.ts
export type Rarity =
  | "REG"
  | "SPE"
  | "RAR"
  | "WBE"
  | "WGK"
  | "MVP"
  | "ATLE"
  | "YS";

export interface CardPoolEntry {
  id: string;
  name: string;
  rarity: Rarity;
  positions: string[];
  version: string;
}